requirements.txt
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
greenlet==3.0.3
idna==3.6
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
parso==0.8.4
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
